05_11중간패치
다음이 마지막이될것같습니다..
가까운 적에게 유도 미사일을
키를 누르면 멀리있는적을 향해 (Linq)사용하게 하였습니다.
공격시 세이더를 통해 외각선이 반짝!였다가 다시 돌아옵니다!
기본적으로 게임시작시 플레이어가 플레이 했던 데이터를 가지고 시작합니다!

앞으로 만들어야하는것

데이터 초기화 : 메뉴에서 버튼을 통해 진행 데이터를 초기화하기
스텟업 버튼 구현 : 스테이지 클리어,보스 처치시 얻는 코인을 통하여 코인사용처 구현!

추가해야되는 요소
현재 스테이지 수 표시
해당 스테이지 보스 처치시 스테이지 증가

+++ Assets/Script/Enemy.cs
@@ -58,7 +58,7 @@ public class Enemy : MonoBehaviour
         target = GameObject.FindGameObjectWithTag("Player").transform;
     }
 
-    public void GetDamaged(int Damage)
+    public void GetDamaged(float Damage)
     {
 
         Hp = Hp - Damage;
